# swagger_server/models/create_driver_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
import re  # noqa: F401,E501
from swagger_server import util


class CreateDriverResponse(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, driver_id: str=None):  # noqa: E501
        """CreateDriverResponse - a model defined in Swagger

        :param driver_id: The driver_id of this CreateDriverResponse.  # noqa: E501
        :type driver_id: str
        """
        self.swagger_types = {
            'driver_id': str
        }

        self.attribute_map = {
            'driver_id': 'driverId'
        }
        self._driver_id = driver_id

    @classmethod
    def from_dict(cls, dikt) -> 'CreateDriverResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CreateDriverResponse of this CreateDriverResponse.  # noqa: E501
        :rtype: CreateDriverResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def driver_id(self) -> str:
        """Gets the driver_id of this CreateDriverResponse.


        :return: The driver_id of this CreateDriverResponse.
        :rtype: str
        """
        return self._driver_id

    @driver_id.setter
    def driver_id(self, driver_id: str):
        """Sets the driver_id of this CreateDriverResponse.


        :param driver_id: The driver_id of this CreateDriverResponse.
        :type driver_id: str
        """
        if driver_id is None:
            raise ValueError("Invalid value for `driver_id`, must not be `None`")  # noqa: E501

        self._driver_id = driver_id