# swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: IPL Tracking - Microservice API
  description: API description for Tracking microservice.
  version: 1.0.0
servers:
- url: http://localhost:8080/tracking
  description: Localhost - BT3 Dev
...
paths:
...
  /drivers:
    get:
      tags:
      - Driver
      summary: Get Drivers list
      description: This operation is used to retrieve a list of Drivers.
      operationId: list_drivers
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListDriversResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Unexpected errors not mapped to status codes mentioned above
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.driver_controller
    post:
      tags:
      - Driver
      summary: Create new Driver
      description: This operation is usedto create a new Driver on System.
      operationId: create_driver
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDriverRequest'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDriverResponse'
        "400":
          description: Parameters missing in request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: There is already a driver registered with same name or mail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Unexpected errors not mapped to status codes mentioned above
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.driver_controller
  /drivers/{driverId}:
    get:
      tags:
      - Driver
      summary: Get a single Driver's info
      description: This operation is used to retrieve the details of a specific device.
      operationId: get_driver
      parameters:
      - name: driverId
        in: path
        description: Unique identifier of the Driver in the database
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/UUID'
        example: 0af401df-7d7a-1f36-817d-7b0a058d0003
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDriverResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Unexpected errors not mapped to status codes mentioned above
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.driver_controller
    delete:
      tags:
      - Driver
      summary: Delete Driver
      description: This operation is delete a Driver.
      operationId: delete_driver
      parameters:
      - name: driverId
        in: path
        description: Unique identifier of the Driver in the database
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/UUID'
        example: 0af401df-7d7a-1f36-817d-7b0a058d0003
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: Delete a driver that has vehicles is not allowed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Unexpected errors not mapped to status codes mentioned above
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.driver_controller
    put:
      tags:
      - Driver
      summary: Update Driver's attributes
      description: This operation is used to update some of the Driver's attributes.
      operationId: update_driver
      parameters:
      - name: driverId
        in: path
        description: Unique identifier of the Driver in the database
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/UUID'
        example: 0af401df-7d7a-1f36-817d-7b0a058d0003
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDriverRequest'
        required: true
      responses:
        "204":
          description: No Content
        "400":
          description: "Parameters missing in request or request is malformed. For\
            \ example, the mail is not in the correct format."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "409":
          description: There is already a Driver registered with same name or email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Unexpected errors not mapped to status codes mentioned above
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.driver_controller
components:
  schemas:
...
    GetDriverResponse:
      required:
      - driverId
      - mail
      - name
      - phone
      type: object
      properties:
        driverId:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          description: Complete Driver name
          example: Foo bar
        customerId:
          $ref: '#/components/schemas/UUID'
        phone:
          type: string
          description: Telephone number
          example: "351912000111"
        mail:
          type: string
          description: E-mail address
          format: email
          example: foo@bar.com
      example:
        driverId: 0af401df-7d7a-1f36-817d-7b0a058d0003
        mail: foo@bar.com
        phone: "351912000111"
        name: Foo bar
        customerId: null
    CreateDriverRequest:
      required:
      - mail
      - name
      - phone
      type: object
      properties:
        name:
          type: string
          description: Complete Driver name
          example: Foo bar
        customerId:
          $ref: '#/components/schemas/UUID'
        phone:
          type: string
          description: Telephone number
          example: "351912000111"
        mail:
          type: string
          description: E-mail address
          format: email
          example: foo@bar.com
    CreateDriverResponse:
      required:
      - driverId
      type: object
      properties:
        driverId:
          $ref: '#/components/schemas/UUID'
      example:
        driverId: 0af401df-7d7a-1f36-817d-7b0a058d0003
    ListDriversResponse:
      required:
      - content
      - totalResults
      type: object
      properties:
        content:
          type: array
          description: list of paged items
          items:
            $ref: '#/components/schemas/GetDriverResponse'
        totalResults:
          type: integer
          description: total number of records
          format: int64
          example: 50
      description: generic paged response
      example:
        totalResults: 50
        content:
        - driverId: 0af401df-7d7a-1f36-817d-7b0a058d0003
          mail: foo@bar.com
          phone: "351912000111"
          name: Foo bar
          customerId: null
        - driverId: 0af401df-7d7a-1f36-817d-7b0a058d0003
          mail: foo@bar.com
          phone: "351912000111"
          name: Foo bar
          customerId: null
    UpdateDriverRequest:
      type: object
      properties:
        name:
          type: string
          description: Complete Driver name
          example: Foo bar
        customerId:
          $ref: '#/components/schemas/UUID'
        phone:
          type: string
          description: Telephone number
          example: "351912000111"
        mail:
          type: string
          description: E-mail address
          format: email
          example: foo@bar.com
...
    ErrorResponse:
      required:
      - errorCode
      - errorMessage
      - errorType
      type: object
      properties:
        code:
          type: string
          description: Application error code
          example: XYZ0001
        type:
          $ref: '#/components/schemas/ErrorTypeEnum'
        message:
          type: string
          description: Short error explanation
          example: Unauthorized
        details:
          type: string
          description: Short error detail
          example: Parameter X has invalid value
    ErrorTypeEnum:
      type: string
      description: Error Type
      enum:
      - PERSISTENCE
      - BUSINESS
      - COMMUNICATION
      - UNKNOWN
    UUID:
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      type: string
      example: 0af401df-7d7a-1f36-817d-7b0a058d0003
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      x-basicInfoFunc: swagger_server.controllers.authorization_controller.check_basicAuth